Moved test code to prepare for more testing
