+++ playground/playground_suite_test.go
@@ -1,17 +1,43 @@
 package playground_test
 
 import (
+	"bytes"
 	"io"
 	"testing"
 
+	"github.com/kkrull/gohttp/httptest"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/types"
 )
 
 func TestPlayground(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "playground")
 }
+
+/* Matchers */
+
+func ShouldHaveNoBody(response *bytes.Buffer, status int, reason string) func() {
+	return func() {
+		responseMessage := httptest.ParseResponse(response)
+		responseMessage.ShouldBeWellFormed()
+		responseMessage.StatusShouldBe(status, reason)
+		responseMessage.HeaderShould("Content-Length", Equal("0"))
+		responseMessage.BodyShould(BeEmpty())
+	}
+}
+
+func ContainSubstrings(values []string) types.GomegaMatcher {
+	valueMatchers := make([]types.GomegaMatcher, len(values))
+	for i, value := range values {
+		valueMatchers[i] = ContainSubstring(value)
+	}
+
+	return SatisfyAll(valueMatchers...)
+}
+
+/* ReadOnlyResourceMock */
 
 type ReadOnlyResourceMock struct {
 	getCalled     bool
@@ -42,6 +68,8 @@ func (mock *ReadOnlyResourceMock) Options(client io.Writer) {
 func (mock *ReadOnlyResourceMock) OptionsShouldHaveBeenCalled() {
 	ExpectWithOffset(1, mock.optionsCalled).To(BeTrue())
 }
+
+/* ReadWriteResourceMock */
 
 type ReadWriteResourceMock struct {
 	getCalled     bool
